Addressing comments, moving the interface to DDB package.
